move find_bugs to its own cli module
